refactor: decompoe Agendamento.vue em View e subcomponentes
- Move a lógica principal e controle de estado para views/AgendamentoView.vue
- Extrai a busca de pacientes para components/agendamento/AgendamentoBusca.vue
- Isola a lógica do calendário em components/agendamento/AgendamentoCalendario.vue
- Cria components/agendamento/AgendamentoResumo.vue para exibição de dados do paciente
- Move o formulário de detalhes para components/agendamento/AgendamentoForm.vue
- Extrai o modal de alertas para components/agendamento/AgendamentoConfirmacaoModal.vue
- Atualiza o roteador para apontar para a nova View

Fixes #44
