+++ frontend/src/components/agendamento/AgendamentoResumo.vue
@@ -1,0 +1,42 @@
+<script lang="ts" setup>
+import {Card, CardContent, CardHeader, CardTitle} from '@/components/ui/card'
+import {Label} from '@/components/ui/label'
+import type {Paciente} from '@/types'
+
+defineProps<{
+  paciente: Paciente
+  protocolo: any
+  ultimoAgendamento: any
+}>()
+</script>
+
+<template>
+  <Card class="bg-blue-50 border-blue-200">
+    <CardHeader>
+      <CardTitle>Resumo do Paciente</CardTitle>
+    </CardHeader>
+    <CardContent class="space-y-3">
+      <div>
+        <p class="font-medium text-lg">{{ paciente.nome }}</p>
+        <p class="text-sm text-gray-600">Reg: {{ paciente.registro }}</p>
+      </div>
+
+      <div v-if="protocolo" class="pt-3 border-t border-blue-200">
+        <Label class="text-xs text-gray-500">Protocolo</Label>
+        <p class="font-medium">{{ protocolo.nome }}</p>
+        <p class="text-sm text-gray-600">{{ protocolo.duracao }} min • Grupo {{ protocolo.grupoInfusao }}</p>
+      </div>
+
+      <div v-if="ultimoAgendamento" class="pt-3 border-t border-blue-200 bg-white/50 p-3 rounded">
+        <Label class="text-xs text-gray-500">Último Agendamento</Label>
+        <p class="text-sm font-medium">{{ new Date(ultimoAgendamento.data).toLocaleDateString('pt-BR') }}</p>
+        <p class="text-sm text-gray-600">Ciclo {{ ultimoAgendamento.cicloAtual }} - {{
+            ultimoAgendamento.diaCiclo
+          }}</p>
+      </div>
+      <div v-else class="pt-3 border-t border-blue-200">
+        <p class="text-sm text-gray-500 italic">Primeiro agendamento</p>
+      </div>
+    </CardContent>
+  </Card>
+</template>
