+++ frontend/src/router/index.ts
@@ -4,7 +4,7 @@ import LoginView from '@/views/LoginView.vue'
 import Layout from '@/components/Layout.vue'
 import PacientesView from '@/views/PacientesView.vue'
 import AgendaView from '@/views/AgendaView.vue'
-import Agendamento from '@/components/Agendamento.vue'
+import AgendamentoView from '@/views/AgendamentoView.vue'
 import Farmacia from '@/components/Farmacia.vue'
 import RelatoriosView from '@/views/RelatoriosView.vue'
 import ProtocolosView from '@/views/ProtocolosView.vue'
@@ -18,7 +18,7 @@ const router = createRouter({
       children: [{path: '', redirect: '/agenda'},
         {path: 'pacientes', name: 'Pacientes', component: PacientesView },
         {path: 'agenda', name: 'Agenda', component: AgendaView},
-        {path: 'agendamento', name: 'Agendamento', component: Agendamento},
+        {path: 'agendamento', name: 'Agendamento', component: AgendamentoView},
         {path: 'farmacia', name: 'Farmacia', component: Farmacia},
         {path: 'relatorios', name: 'Relatorios', component: RelatoriosView},
         {path: 'protocolos', name: 'Protocolos', component: ProtocolosView},
